+++ src/main/java/org/alexherr/exchangerateservice/model/ConvertedCurrency.java
@@ -1,15 +1,16 @@
 package org.alexherr.exchangerateservice.model;
 
+import java.math.BigDecimal;
 import java.time.LocalDate;
 
 public class ConvertedCurrency {
 
     private LocalDate date;
     private String currencyShortName;
-    private float result;
+    private BigDecimal result;
     private String base;
 
-    public ConvertedCurrency(LocalDate date, String currencyShortName, float result, String base) {
+    public ConvertedCurrency(LocalDate date, String currencyShortName, BigDecimal result, String base) {
         this.date = date;
         this.currencyShortName = currencyShortName;
         this.result = result;
@@ -24,7 +25,7 @@ public class ConvertedCurrency {
         return currencyShortName;
     }
 
-    public float getResult() {
+    public BigDecimal getResult() {
         return result;
     }
